Handle HTTP code, and decrement songNum if failed
In bot.js, tested Promises instead of callback to handle errors or
responses from M.post. Seems to work but cannot catch
Uncaught Error getaddrinfo ENOTFOUND
I think it is a bug in the mastodon-api library I'm using.
https://stackoverflow.com/questions/64283656/nodejs-getaddrinfo-enotfound-uncaught

Unfortunately, if the song does not get tooted because of this error,
the songNumber was incremented before the failure.